# Notes-Server/swagger/swagger.json
{
  "openapi" : "3.0.0",
  "servers" : [ {
    "url" : "http://localhost:3000/api/v1"
  } ],
  "info" : {
    "description" : "Add, edit, manager and share notes between users",
    "version" : "0.1-oas3",
    "title" : "Notes"
  },
...
  "paths" : {
...
    "/api/v1/remainders/create" : {
      "post" : {
        "tags" : [ "remainders" ],
        "summary" : "Create a new remainder",
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Remainder"
                }
              }
            }
          },
          "403" : {
            "description" : "Not authenticated",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          }
        },
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "properties" : {
                  "userid" : {
                    "type" : "string"
                  },
                  "noteid" : {
                    "type" : "string"
                  },
                  "remaindertime" : {
                    "type" : "string"
                  }
                },
                "type" : "object"
              }
            }
          },
          "required" : true
        }
      }
    },
    "/api/v1/remainders/delete" : {
      "post" : {
        "tags" : [ "remainders" ],
        "summary" : "Delete a remainder",
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "403" : {
            "description" : "Not authenticated",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "404" : {
            "description" : "Cannot find remainder/Cannot find note",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          }
        },
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "properties" : {
                  "userid" : {
                    "type" : "string"
                  },
                  "noteid" : {
                    "type" : "string"
                  },
                  "remainderid" : {
                    "type" : "string"
                  }
                },
                "type" : "object"
              }
            }
          },
          "required" : true
        }
      }
    },
    "/api/v1/remainders/dismiss" : {
      "post" : {
        "tags" : [ "remainders" ],
        "summary" : "Dismiss a remainder",
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "403" : {
            "description" : "Not authenticated",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "404" : {
            "description" : "Cannot find remainder",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          }
        },
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "properties" : {
                  "userid" : {
                    "type" : "string"
                  },
                  "remainderid" : {
                    "type" : "string"
                  }
                },
                "type" : "object"
              }
            }
          },
          "required" : true
        }
      }
    },
    "/api/v1/remainders/snooze" : {
      "post" : {
        "tags" : [ "remainders" ],
        "summary" : "Snooze a remainder for a given time",
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "403" : {
            "description" : "Not authenticated",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "404" : {
            "description" : "Cannot find remainder",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          }
        },
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "properties" : {
                  "userid" : {
                    "type" : "string"
                  },
                  "remainderid" : {
                    "type" : "string"
                  },
                  "snoozetime" : {
                    "type" : "number"
                  }
                },
                "type" : "object"
              }
            }
          },
          "required" : true
        }
      }
    },
    "/api/v1/remainders/" : {
      "post" : {
        "tags" : [ "remainders" ],
        "summary" : "Get all remainders created by a user",
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "responses" : {
          "200" : {
            "description" : "success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Remainder"
                  }
                }
              }
            }
          },
          "403" : {
            "description" : "Not authenticated",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "404" : {
            "description" : "No remainders found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          }
        },
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "properties" : {
                  "userid" : {
                    "type" : "string"
                  }
                },
                "type" : "object"
              }
            }
          },
          "required" : true
        }
      }
    },
    "/api/v1/remainders/detail" : {
      "post" : {
        "tags" : [ "remainders" ],
        "summary" : "Get details of a particular remainder",
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "responses" : {
          "200" : {
            "description" : "success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Remainder"
                }
              }
            }
          },
          "403" : {
            "description" : "Not authenticated",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "404" : {
            "description" : "Cannot find remainder",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          }
        },
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "properties" : {
                  "userid" : {
                    "type" : "string"
                  },
                  "remainderid" : {
                    "type" : "string"
                  }
                },
                "type" : "object"
              }
            }
          },
          "required" : true
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "Response" : {
        "properties" : {
          "message" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "Remainder" : {
        "properties" : {
          "remainderId" : {
            "type" : "string"
          },
          "userId" : {
            "type" : "string"
          },
          "noteId" : {
            "type" : "string"
          },
          "remainderTime" : {
            "type" : "string"
          },
          "isDismissed" : {
            "type" : "boolean"
          },
          "createdOn" : {
            "type" : "string"
          }
        }
      }
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}